Use Float32 for nix.DataType.Float
Was np.float_ which is equivalent to double (float64)

+++ nixio/datatype.py
@@ -24,7 +24,7 @@ class DataType(object):
     Int16 = np.int16
     Int32 = np.int32
     Int64 = np.int64
-    Float = np.float_
+    Float = np.float32
     Double = np.double
     String = np.unicode_
     Bool = np.bool_
